+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
   - [583](https://github.com/thoth-pub/thoth/issues/583) - Add new field, Permanently Withdrawn Date, to Work for Out-of-print or Withdrawn from Sale Works.
 
+### Changed
+ - [218](https://github.com/thoth-pub/thoth/issues/218) - Make series ISSN optional
+
 ## [[0.12.2]](https://github.com/thoth-pub/thoth/releases/tag/v0.12.2) - 2024-04-16
 ### Added
   - [581](https://github.com/thoth-pub/thoth/issues/581) - Add crossmark policy DOI to imprint record